requirements.txt
streamlit
tensorflow
Pillow
numpy
huggingface_hub
